- Api service was refactored

+++ BGLibrary/BGNet.TestAssignment.Api/Controllers/AuthorController.cs
@@ -1,8 +1,9 @@
 ﻿using BGNet.TestAssignment.BusinessLogic.Services;
-using BGNet.TestAssignment.DataAccess.Entities;
 using BGNet.TestAssignment.Models.Dtos;
+using BGNet.TestAssignment.Models.Responses;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Net;
 
 namespace BGNet.TestAssignment.Api.Controllers;
 
@@ -23,37 +24,92 @@ public class AuthorController : ControllerBase
     [HttpGet]
     public IActionResult GetList()
     {
-        return Ok(_authorRepository.GetAll());
+        return Ok(new ApiResponse<IEnumerable<AuthorDto>>
+        {
+            StatusCode = (int)HttpStatusCode.OK,
+            Data = _authorRepository.GetAll(),
+            Message = "Success",
+        });
     }
 
     [HttpGet("{id:int}")]
     public IActionResult GetAuthor(int id)
     {
-        return Ok(_authorRepository.GetById(id));
+        IActionResult result;
+
+        var author = _authorRepository.GetById(id);
+
+        if (author is not null)
+        {
+            result = Ok(new ApiResponse<AuthorDto>
+            {
+                StatusCode = (int)HttpStatusCode.OK,
+                Data = _authorRepository.GetById(id),
+                Message = "Success",
+            });
+        }
+        else
+        {
+            result = BadRequest(new ApiResponse
+            {
+                StatusCode = (int)HttpStatusCode.OK,
+                Errors = new[] { $"Author with id {id} not found" },
+            });
+        }
+
+        return result;
     }
 
     [HttpPost]
     public IActionResult Create(AuthorDto author)
     {
-        var createdAuthor = _authorRepository.Create(author);
-
-        return Created("Success", createdAuthor.Id);
-    }   
+        return CreatedAtAction(nameof(Create), new ApiResponse<AuthorDto>
+        {
+            StatusCode = (int)HttpStatusCode.Created,
+            Data = _authorRepository.Create(author),
+            Message = "Success",
+        });
+    }
 
     [HttpPut]
     public IActionResult Update(AuthorDto author)
     {
         _authorRepository.Update(author);
 
-        return Ok("Success");
+        return Ok(new ApiResponse
+        {
+            StatusCode = (int)HttpStatusCode.OK,
+            Message = "Success",
+        });
     }
 
     [HttpDelete("{id:int}")]
     public IActionResult Delete(int id)
     {
-        _authorRepository.Delete(id);
+        IActionResult result;
 
-        return Ok("Success");
+        var authorToDelete = _authorRepository.GetById(id);
+
+        if (authorToDelete is not null)
+        {
+            _authorRepository.Delete(id);
+
+            result = Ok(new ApiResponse
+            {
+                StatusCode = (int)HttpStatusCode.OK,
+                Message = $"Author with id {id} was deleted successfully",
+            });
+        }
+        else
+        {
+            result = BadRequest(new ApiResponse
+            {
+                StatusCode = (int)HttpStatusCode.BadRequest,
+                Errors = new[] { $"Author with id {id} not found" },
+            });
+        }
+
+        return result;
     }
 
     #endregion
